+++ lib/src/widgets/video_ad_widget.dart
@@ -1,2 +1,2 @@
-export 'video_ad_widget_mobile.dart'
-    if (dart.library.html) 'video_ad_widget_web.dart';
+export 'video_ad_widget_web.dart'
+    if (dart.library.io) 'video_ad_widget_mobile.dart';